api/serializers.py
from rest_framework import serializers
from orders.models import Order


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = ('order_number', 'usd_price', 'rub_price', 'shipment_date')
